[resources] Add Documentation around new writable/readable flags (#21)
* Make `SomeResource.find` docs/api a bit clearer

A concern was raised around what `find` returns in https://github.com/graphiti-api/graphiti/issues/170. The change here calls out explicitly what `find` returns and how to access to model proper.

* Update resource flags for reading and writing

A recent PR granted Graphiti the capability to pass symbols, strings, and blocks in addition to booleans for the `readable` and `writable` flags. This allows clients to more dynamically evaluate if a specific resource should have access based on permissions or other, more dynamic conditions.

* Update resource flags for reading and writing

A recent PR granted Graphiti the capability to pass symbols, strings, and blocks in addition to booleans for the `readable` and `writable` flags. This allows clients to more dynamically evaluate if a specific resource should have access based on permissions or other, more dynamic conditions.